Define all ServiceModel extensions inside the same ServiceModelExt object
This allows us to use the same name `toApi` for all the extensions
