--- modules/api/src/main/scala/ca/rmen/poetassistant/api/model/mapping/SyllableRhymesServiceModelExt.scala
@@ -1,33 +1,0 @@
-/*
- * Copyright (c) 2021 - present Carmen Alvarez
- *
- * This file is part of Poet Assistant.
- *
- * Poet Assistant is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Poet Assistant is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with Poet Assistant.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package ca.rmen.poetassistant.api.model.mapping
-
-import ca.rmen.poetassistant.api.model.SyllableRhymesApiModel
-import ca.rmen.poetassistant.service.model.SyllableRhymesServiceModel
-
-object SyllableRhymesServiceModelExt {
-  extension (x: SyllableRhymesServiceModel) def toSyllableRhymesApiModel = SyllableRhymesApiModel(
-    syllables = x.syllables,
-    rhymes = x.rhymes
-  )
-  extension (x: Option[SyllableRhymesServiceModel]) def toOptionSyllableRhymesApiModel: Option[SyllableRhymesApiModel] = {
-    x.map(_.toSyllableRhymesApiModel)
-  }
-}
